reorganized the service entity UI to move common info into the top (always visible) section
