+++ manager/ui/war/src/main/java/io/apiman/manager/ui/client/local/pages/ServiceImplPage.java
@@ -200,6 +200,7 @@ public class ServiceImplPage extends AbstractServicePage {
             public void onSuccess(Void response) {
                 saveButton.onActionComplete();
                 saveButton.setEnabled(false);
+                refreshServiceVersion();
             }
             @Override
             public void onError(Throwable error) {
