+++ manager/ui/war/src/main/java/io/apiman/manager/ui/client/local/site/service-activity.html
@@ -52,28 +52,69 @@
         </div>
       </div>
       <!-- Entity Summary Row -->
-      <div class="row">
-        <div class="col-md-12 apiman-entity-breadcrumb">
-          <div class="title container-fluid">
-            <i class="breadcrumb-icon fa fa-puzzle-piece"></i>
-            <div class="entity emphasis">
-              <a data-field="serviceName" href="service-overview.html" data-role="dummy">dtgov-deploy</a>
-            </div>
-            <div class="versions">
-              <div class="btn-group apiman-entity-action" data-role="dummy" data-field="versions">
-                <button type="button" class="btn btn-default dropdown-toggle" data-toggle="dropdown">
-                  Version: 3.0.1 <span class="caret"></span>
-                </button>
-                <ul class="dropdown-menu" role="menu">
-                  <li><a href="#">3.0</a></li>
-                  <li><a href="#">2.1.1</a></li>
-                  <li><a href="#">2.1</a></li>
-                  <li><a href="#">2.0.1</a></li>
-                  <li><a href="#">2.0</a></li>
-                  <li><a href="#">1.0</a></li>
-                </ul>
+      <div class="apiman-entity-summary">
+        <div class="row apiman-entity-breadcrumb">
+          <div class="col-md-12">
+            <div class="title container-fluid">
+              <i class="breadcrumb-icon fa fa-puzzle-piece"></i>
+              <div class="entity emphasis">
+                <a data-field="serviceName" href="service-overview.html" data-role="dummy">dtgov-deploy</a>
               </div>
-              <a data-permission="svcEdit" href="new-serviceversion.html" class="btn btn-primary apiman-entity-action" data-field="toNewServiceVersion" data-i18n-key="new-version">New Version</a>
+              <div class="versions">
+                <div class="btn-group apiman-entity-action" data-role="dummy" data-field="versions">
+                  <button type="button" class="btn btn-default dropdown-toggle" data-toggle="dropdown">
+                    Version: 3.0.1 <span class="caret"></span>
+                  </button>
+                  <ul class="dropdown-menu" role="menu">
+                    <li><a href="#">3.0</a></li>
+                    <li><a href="#">2.1.1</a></li>
+                    <li><a href="#">2.1</a></li>
+                    <li><a href="#">2.0.1</a></li>
+                    <li><a href="#">2.0</a></li>
+                    <li><a href="#">1.0</a></li>
+                  </ul>
+                </div>
+                <a data-permission="svcEdit" href="new-serviceversion.html" class="btn btn-primary apiman-entity-action" data-field="toNewServiceVersion" data-i18n-key="new-version">New Version</a>
+              </div>
+            </div>
+            <hr />
+          </div>
+        </div>
+        <div class="row apiman-entity-metadata">
+          <div class="col-md-7" style="margin-bottom: 8px">
+            <!-- Service Summary -->
+            <div class="description" data-field="description" data-role="dummy">The API that allows dtgov governance users and workflows to deploy artifacts to a configured runtime environment.</div>
+            <div style="padding: 8px">
+              <div class="entity-info-with-icon">
+                <i class="fa fa-clock-o fa-fw"></i>
+                <span class="apiman-label-faded" data-i18n-key="created-on">Created on</span>
+                <span data-field="createdOn" data-role="dummy" >Feb 3, 2011</span>
+              </div>
+              <div class="entity-info-with-icon">
+                <i class="fa fa-user fa-fw"></i>
+                <span class="apiman-label-faded" data-i18n-key="created-by">Created by</span>
+                <span><a href="user-orgs.html" data-field="createdBy" data-role="dummy">Clark Kent</a></span>
+              </div>
+            </div>
+            <div class="entity-info-with-icon">
+              <span data-i18n-key="status-label">Status:</span>
+              <span class="apiman-label apiman-label-warning" data-field="status" data-role="dummy">Ready</span>
+            </div>
+          </div>
+          <div class="col-md-5" data-permission="svcEdit">
+            <div>
+              <div data-status="Published">
+                <a data-i18n-key="ttdo-new-svc-contract" data-field="ttd_toNewContract" href="new-contract.html">Link my Application to this Service (New Contract)</a>
+              </div>
+              <div>
+                <a data-i18n-key="ttdo-new-svc-version" data-field="ttd_toNewServiceVersion" href="new-serviceversion.html">Create a new version of this Service (New Version)</a>
+              </div>
+            </div>
+            <!-- The Publish Action -->
+            <div class="apiman-divider-40"></div>
+            <div class="">
+              <button class="btn btn-primary" data-field="publishButton" data-i18n-key="publish" placeholder="Publishing..." data-icon="fa-cog">Publish</button>
+              <button class="btn btn-warning" data-field="retireButton" data-i18n-key="retire" placeholder="Retiring..." data-icon="fa-cog">Retire</button>
             </div>
           </div>
         </div>
@@ -89,9 +130,9 @@
             <li><a data-field="toServiceImpl" href="service-impl.html" data-i18n-key="implementation">Implementation</a></li>
             <li class=""><a data-field="toServicePlans" href="service-plans.html" data-i18n-key="plans">Plans</a></li>
             <li class=""><a data-field="toServicePolicies" href="service-policies.html" data-i18n-key="policies">Policies</a></li>
-            <li class="north"><a data-field="toServiceContracts" href="service-contracts.html" data-i18n-key="contracts">Contracts</a></li>
+            <li class=""><a data-field="toServiceContracts" href="service-contracts.html" data-i18n-key="contracts">Contracts</a></li>
             <li class="active"><a data-field="toServiceActivity" href="service-activity.html" data-i18n-key="activity">Activity</a></li>
-            <li class="south last"><a data-role="dummy">&nbsp;</a></li>
+            <li class="last"><a data-role="dummy">&nbsp;</a></li>
           </ul>
         </div>
         <!-- /Left hand nav -->
