+++ manager/ui/war/src/main/webapp/index.html
@@ -8,8 +8,8 @@
 
     <link href="bootstrap-3.1.1/css/bootstrap.min.css" rel="stylesheet" media="screen"></link>
     <link href="patternfly-1.0/css/patternfly.min.css" rel="stylesheet" media="screen"></link>
-    <link href="css/apiman.css" rel="stylesheet"></link>
-    <link href="css/apiman-responsive.css" rel="stylesheet"></link>
+    <link href="css/apiman.css?iter=1" rel="stylesheet"></link>
+    <link href="css/apiman-responsive.css?iter=1" rel="stylesheet"></link>
     <link href="css/branding.css" rel="stylesheet"></link>
 
     <script type="text/javascript">
